requirements.txt
Pillow==10.4.0
pandas==2.2.2
numpy==1.26.4
